feat: virtual dts per framework
