+++ examples/preact-router/src/vite-env.d.ts
@@ -1,2 +1,3 @@
 /// <reference types="vite/client" />
-/// <reference types="vite-plugin-pwa/client" />
+/// <reference types="vite-plugin-pwa/preact" />
+/// <reference types="vite-plugin-pwa/info" />
