+++ examples/preact-router/tsconfig.json
@@ -17,7 +17,7 @@
     "jsx": "preserve",
     "jsxFactory": "h",
     "jsxFragmentFactory": "Fragment",
-    "types": ["vite/client", "vite-plugin-pwa", "vite-plugin-pwa/client"]
+    "types": ["vite/client", "vite-plugin-pwa/preact", "vite-plugin-pwa/info"]
   },
   "include": ["src"]
 }
