+++ examples/react-router/src/vite-env.d.ts
@@ -1,2 +1,3 @@
 /// <reference types="vite/client" />
-/// <reference types="vite-plugin-pwa/client" />
+/// <reference types="vite-plugin-pwa/react" />
+/// <reference types="vite-plugin-pwa/info" />
