+++ examples/react-router/tsconfig.json
@@ -15,8 +15,9 @@
     "noEmit": true,
     "jsx": "react",
     "types": [
-      "vite-plugin-pwa",
-      "vite-plugin-pwa/client"
+      "vite/client",
+      "vite-plugin-pwa/react",
+      "vite-plugin-pwa/info"
     ]
   },
   "include": ["./src"]
