+++ examples/solid-router/src/ReloadPrompt.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable react/no-unknown-property */
 import type { Component } from 'solid-js'
 import { Show } from 'solid-js'
 import { useRegisterSW } from 'virtual:pwa-register/solid'
