+++ examples/solid-router/tsconfig.json
@@ -8,6 +8,6 @@
     "esModuleInterop": true,
     "jsx": "preserve",
     "jsxImportSource": "solid-js",
-    "types": ["vite/client", "vite-plugin-pwa", "vite-plugin-pwa/client"]
+    "types": ["vite/client", "vite-plugin-pwa/info", "vite-plugin-pwa/solid"]
   }
 }
