+++ examples/svelte-routify/tsconfig.json
@@ -14,9 +14,6 @@
       */
     "allowJs": true,
     "checkJs": true,
-    "types": [
-      "vite-plugin-pwa/client"
-    ]
   },
   "include": ["src/**/*.d.ts", "src/**/*.ts", "src/**/*.js", "src/**/*.svelte"]
 }
