+++ examples/svelte-routify/vite-env.d.ts
@@ -1,3 +1,4 @@
 /// <reference types="svelte" />
 /// <reference types="vite/client" />
-/// <reference types="vite-plugin-pwa/client" />
+/// <reference types="vite-plugin-pwa/svelte" />
+/// <reference types="vite-plugin-pwa/info" />
