+++ examples/vanilla-ts-dev-options/src/vite-env.ts
@@ -1,3 +1,3 @@
 /// <reference types="vite/client" />
-/// <reference types="vite-plugin-pwa/client" />
+/// <reference types="vite-plugin-pwa/vanillajs" />
 /// <reference types="vite-plugin-pwa/info" />
