+++ examples/vanilla-ts-dev-options/tsconfig.json
@@ -15,7 +15,7 @@
     "noUnusedParameters": true,
     "noImplicitReturns": true,
     "skipLibCheck": true,
-    "types": ["vite-plugin-pwa"]
+    "types": ["vite-plugin-pwa/vanillajs", "vite-plugin-pwa/info"]
   },
   "include": ["src"]
 }
