+++ examples/vanilla-ts-no-ip/tsconfig.json
@@ -15,7 +15,7 @@
     "noUnusedParameters": true,
     "noImplicitReturns": true,
     "skipLibCheck": true,
-    "types": ["vite-plugin-pwa", "vite-plugin-pwa/client", "vite-plugin-pwa/info"]
+    "types": ["vite/client", "vite-plugin-pwa/client", "vite-plugin-pwa/info"]
   },
   "include": ["src/**/*.ts"]
 }
